Remove old worker subtypes
