--- src/roles/upgrader.ts
@@ -1,7 +1,0 @@
-import { WorkerCreep } from '../virtualCreeps/workerCreep';
-
-export class Upgrader extends WorkerCreep {
-    protected performDuties() {
-        this.runUpgradeJob();
-    }
-}
